Factor out table class
