+++ pull_scan_metadata/src/python/scan_metadata_app/main.py
@@ -3,82 +3,14 @@
 import logging
 import re
 from io import BytesIO
-from typing import Dict, List, Optional
+from typing import Dict, List
 
 import pandas as pd
 from flywheel import FileSpec, Group, Project
 from flywheel_adaptor.flywheel_proxy import FlywheelProxy
+from tabular_data.site_table import SiteTable
 
 log = logging.getLogger(__name__)
-
-
-class SCANTable:
-    """Wrapper for data frame for SCAN metadata table.
-
-    Supports splitting table by Center ID. Center ID could be in column
-    named ADCID or SITE.
-    """
-
-    def __init__(self, *, data: pd.DataFrame, site_id_column: str) -> None:
-        self.__data_table = data
-        self.__site_column = site_id_column
-
-    @classmethod
-    def create_from(cls, table_data: pd.DataFrame) -> Optional['SCANTable']:
-        """Recognizes which column is used for site ID.
-
-        Args:
-          table_data: the data frame with data
-        Returns:
-          a wrapper object for the data frame or None if no center id column
-        """
-        if 'ADCID' in table_data.columns:
-            site_id_name = 'ADCID'
-        elif 'SITE' in table_data.columns:
-            site_id_name = 'SITE'
-        else:
-            return None
-
-        return SCANTable(data=table_data, site_id_column=site_id_name)
-
-    def site_keys(self) -> Dict[str, Optional[str]]:
-        """Returns a map from site column values to the ADCID.
-
-        Note: it is possible that an ADCID may not be found, in which case
-        returns None.
-
-        Returns:
-          dictionary mapping from site column value to ADCID or None
-        """
-        site_ids = self.__data_table.loc(self.__site_column)
-        return {site_key: self.get_adcid(site_key) for site_key in site_ids}
-
-    def get_adcid(self, key) -> Optional[str]:
-        """Returns the ADCID for the site column ID value.
-
-        Args:
-          key: the column key value
-        Returns:
-          the corresponding ADCID or None if none
-        """
-        if self.__site_column == 'ADCID':
-            return key
-
-        match = re.search(r"([^(]+)\(ADC\s?(\d+)\)", key)
-        if not match:
-            return None
-        return match.group(2).strip()
-
-    def select_site(self, site_id) -> pd.DataFrame:
-        """Selects the rows of the table for the site.
-
-        Args:
-          site_id: the ID of the table to select
-        Returns:
-          Data frame with rows of the table for the site_id
-        """
-        return self.__data_table.loc[self.__data_table[self.__site_column] ==
-                                     site_id]
 
 
 def read_data(*, s3_client, bucket_name: str, file_name: str):
@@ -130,23 +62,58 @@ def build_center_map(*, proxy: FlywheelProxy,
     return center_map
 
 
-def upload_file(*, project: Project, site_table: pd.DataFrame,
-                file_name: str) -> None:
+def split_table(*, proxy: FlywheelProxy, table: SiteTable,
+                center_map: Dict[str, Group], filename: str,
+                destination_label: str) -> None:
+    """Splits the table by center ID and uploads to matching projects in FW.
+
+    Args:
+      proxy: the proxy object for Flywheel instance
+      table: the table to split
+      center_map: the map from adcid tags to center groups
+      filename: the name of the file to use
+      destination_label: the name of the project in center group
+    """
+    site_map = table.site_keys()
+    for site_key, adcid in site_map.items():
+        if not adcid:
+            log.warning('Site %s has no matching ADCID', site_key)
+            continue
+
+        site_table = table.select_site(site_key)
+        if not site_table:
+            log.error('Unable to select site data for ADCID %s', adcid)
+            continue
+
+        center_group = center_map.get(f'adcid-{adcid}')
+        if not center_group:
+            log.error('Did not find group for ADCID %s', adcid)
+            continue
+
+        project = proxy.get_project(group=center_group,
+                                    project_label=destination_label)
+        if not project:
+            log.error('Unable to access project %s', destination_label)
+            continue
+
+        log.info("Uploading file %s for adcid %s", filename, adcid)
+        upload_file(project=project, site_table=site_table, file_name=filename)
+
+
+def upload_file(*, project: Project, site_table: str, file_name: str) -> None:
     """Creates CSV file and uploads to Flywheel project.
 
     Args:
       project: the Flywheel project
-      site_table: the table for site data
+      site_table: the CSV site data
       file_name: the name of the file on Flywheel
     """
-    csv_data = site_table.to_csv(index=False)
     file_spec = FileSpec(name=file_name,
-                         contents=csv_data,
+                         contents=site_table,
                          content_type='text/csv')
     project.upload_file(file_spec)
 
 
-# pylint: disable=(too-many-locals)
 def run(*, proxy: FlywheelProxy, table_list: List[str], s3_client,
         center_tag_pattern: str, bucket_name: str,
         destination_label: str) -> None:
@@ -179,7 +146,7 @@ def run(*, proxy: FlywheelProxy, table_list: List[str], s3_client,
             log.error('Unable to access file %s: %s', filename, obj_error)
             continue
 
-        table = SCANTable.create_from(load_table(data))
+        table = SiteTable.create_from(load_table(data))
         if not table:
             log.error(
                 'Table %s does not have a column with recognized center ID',
@@ -187,24 +154,8 @@ def run(*, proxy: FlywheelProxy, table_list: List[str], s3_client,
             continue
 
         log.info("Splitting table %s", table_name)
-        site_map = table.site_keys()
-        for site_key, adcid in site_map.items():
-            if not adcid:
-                log.warning('Site %s has no matching ADCID', site_key)
-                continue
-
-            site_table = table.select_site(site_key)
-            center_group = center_map.get(f'adcid-{adcid}')
-            if not center_group:
-                log.error('Did not find group for ADCID %s', adcid)
-                continue
-
-            project = proxy.get_project(group=center_group,
-                                        project_label=destination_label)
-            if not project:
-                log.error('Unable to access project %s', destination_label)
-                continue
-
-            upload_file(project=project,
-                        site_table=site_table,
-                        file_name=filename)
+        split_table(proxy=proxy,
+                    table=table,
+                    center_map=center_map,
+                    filename=filename,
+                    destination_label=destination_label)
